+++ src/guv_calcs/__init__.py
@@ -1,17 +1,20 @@
-from .room import Room
+from .room import Room, DEFAULT_DIMS
 from .project import Project
 from .lamp import Lamp, LampType, GUVType, LampPlacer, Spectrum, sum_spectrum
 from .calc_zone import CalcVol, CalcPlane
 from .geometry import Polygon2D, to_polar, to_cartesian, attitude
+from .geometry import PlaneGrid, VolGrid
 from .io import get_spectral_weightings
 from .efficacy import InactivationData
 from .safety import PhotStandard, get_tlvs, get_max_irradiance, get_seconds_to_tlv
 from .units import convert_units, convert_length, convert_time
+from .standard_zones import WHOLE_ROOM_FLUENCE, EYE_LIMITS, SKIN_LIMITS
 from ._read import read_export_file, file_to_zone
 from ._version import __version__
 
 __all__ = [
     "Room",
+    "DEFAULT_DIMS",
     "Project",
     "Lamp",
     "LampType",
@@ -21,6 +24,8 @@ __all__ = [
     "sum_spectrum",
     "CalcVol",
     "CalcPlane",
+    "PlaneGrid",
+    "VolGrid",
     "Polygon2D",
     "to_polar",
     "to_cartesian",
@@ -34,6 +39,9 @@ __all__ = [
     "convert_units",
     "convert_length",
     "convert_time",
+    "WHOLE_ROOM_FLUENCE",
+    "EYE_LIMITS",
+    "SKIN_LIMITS",
     "read_export_file",
     "file_to_zone",
 ]